refactor: mypy 44→0 全清零 + ruff 全过（按 8 类顺序清理）
按规则类别逐类清理剩余 mypy（每类验证后进下一类）：

1. _audit_write arg-type ×6：duration_ms: int→float（调用传 elapsed*1000）
2. misc unpack/deleted-e ×6：_parse_holdings_html 签名 4→2 元（与 return 一致）/
   _async_fetch_holdings_one 签名 3→4 元（补 failed）/ macro for e→item（避 except e 同名）
3. assignment+return-value：nav _save_nav_batch 重构局部变量（消除 attr-defined+return）/
   macro result dict[str,object] / domain zh_regime or "" / valuation zip 改名避 shadow /
   dashboard baseline 50→50.0 / model save_path 接 Path
4. arg-type 精准：recommend score/combo or 0.0 / ga _encode 接 RankingConfig /
   fetchers raise last_error or RuntimeError / runner slot or "" / domain regime_label or ""
5. 库签名 type:ignore（6 处，非代码问题）：FastAPI lifespan / OpenAI messages Param /
   LoggerAdapter.log override / sys.stdout.reconfigure / backtest cfg 覆盖 /
   charts float or 0（mypy 不 narrow 过滤）

mypy 78→0（Success: no issues found in 44 source files）
ruff All checks passed
测试：495 passed

+++ app/llm/client.py
@@ -35,7 +35,7 @@ _LLM_TIMEOUT = 300.0           # 单次请求总超时（秒）：大 max_tokens
 
 
 def _audit_write(caller: str, prompt: str, raw_output: str, parsed_result: Any, ok: bool,
-                 duration_ms: int, tokens: int) -> None:
+                 duration_ms: float, tokens: int) -> None:
     """写入 LLM 决策审计（P0-3）：prompt 快照 + 原始输出 + 解析结果，可复现排查。
 
     审计写入失败不阻断主流程（技术记录，容错丢弃）；滚动保留最近 _AUDIT_MAX_ROWS 条。
@@ -120,7 +120,7 @@ def _call_llm(
         try:
             resp = client.chat.completions.create(
                 model=model,
-                messages=messages,
+                messages=messages,  # type: ignore[arg-type]  # OpenAI SDK Param 类型严格，dict 消息运行时兼容
                 temperature=temperature,
                 max_tokens=max_tokens,
             )
